Updates CLI and logging
Use click for cli.
If installed with `pip install --editable .` peddy is run by using command `peddy`.
Adds colored logging instead of prints.
